+++ src/Ray.h
@@ -12,7 +12,7 @@ struct Ray
 
 	bool interceptsWith(const Triangle& t, Vector3* interception) const;
 	bool interceptsWith(const Triangle& t, Vector3* interception, Vector3* normal) const;
-	bool interceptsWith(const Sphere& s, double* firstInterception, double* secondInterception) const;
+	bool interceptsWith(const Sphere& s, float* firstInterception, float* secondInterception) const;
 	float getReflectionAngle(const Sphere& s, const Vector3& p);
 
 };
